Decouple orchestrator resources into separate rest-api definitions
