+++ configserver/src/main/resources/configserver-app/services.xml
@@ -63,8 +63,30 @@
     <component id="com.yahoo.vespa.orchestrator.controller.RetryingClusterControllerClientFactory" bundle="orchestrator" />
     <component id="com.yahoo.vespa.orchestrator.OrchestratorImpl" bundle="orchestrator" />
     
-    <rest-api path="orchestrator" jersey2="true">      
-      <components bundle="orchestrator" />      
+    <rest-api path="orchestrator/v1/suspensions/applications" jersey2="true">
+      <components bundle="orchestrator">
+        <package>com.yahoo.vespa.orchestrator.resources.appsuspension</package>
+      </components>
+    </rest-api>
+    <rest-api path="orchestrator/v1/health" jersey2="true">
+      <components bundle="orchestrator">
+        <package>com.yahoo.vespa.orchestrator.resources.health</package>
+      </components>
+    </rest-api>
+    <rest-api path="orchestrator/v1/hosts" jersey2="true">
+      <components bundle="orchestrator">
+        <package>com.yahoo.vespa.orchestrator.resources.host</package>
+      </components>
+    </rest-api>
+    <rest-api path="orchestrator/v1/suspensions/hosts" jersey2="true">
+      <components bundle="orchestrator">
+        <package>com.yahoo.vespa.orchestrator.resources.hostsuspension</package>
+      </components>
+    </rest-api>
+    <rest-api path="orchestrator/v1/instances" jersey2="true">
+      <components bundle="orchestrator">
+        <package>com.yahoo.vespa.orchestrator.resources.instance</package>
+      </components>
     </rest-api>
 
     <handler id="com.yahoo.vespa.serviceview.StateRequestHandler" bundle="configserver">
